Add high-level documentation to RAML (#34)
I am using the facilities explored in
https://issues.folio.org/browse/FOLIO-1833
to include a Markdown document into the documentation generated from
the RAML and JSON Schemas. As a result, the auto-generated
documentation at
https://s3.amazonaws.com/foliodocs/api/mod-courses/courses.html
and
https://s3.amazonaws.com/foliodocs/api/mod-courses/p/courses.html
should include the overview prose.

I have also removed `MODEL.md` from the top level of this repository,
as it is now significantly out of date and misleading. The new
overview documentation supersedes and updates it.