Simplify reserves records to be added to database (#72)
Previously we were adding complete records with an `itemId`,
`copiedItem`, etc. All we need to POST is the `id`, `courseListingId`
and `copiedItem.barcode`, and the course-reserves module's logic will
pull in the item.

I am not 100% sure that these records will do what we want, because I
have not been able to load them against an empty database. But I think
they will, and we will find out on tonight's rebuild.